Move readColonFile to separate file
Plan9 operating system stores users and groups information in
/adm/users file. This file is similar in strucuture to unix
based /etc/passwd and /etc/group. Since readColonFile resides in
lookup_unix.go file, extracting it makes it possible to reuse the
functionality across all os targets that can can utilize the function.
